#24: Fix failing test.java build target.
hxjava appears to have arbitrarily changed the rules it uses to decide
the name of the output jar. It used to be the same name as that of the
output directory, now it appears to be the same name as that of the
main class.

Annoyingly there is no way to explicitly specify the output jar name.